Refactor ProviderType calls from LookupService
Consolodating them all in ProviderTypeService makes further enhancements
cleaner.
