+++ psm-app/cms-business-process/src/main/java/gov/medicaid/services/impl/LookupServiceBean.java
@@ -16,12 +16,9 @@
 
 package gov.medicaid.services.impl;
 
-import gov.medicaid.domain.model.ApplicantType;
-import gov.medicaid.entities.AgreementDocument;
 import gov.medicaid.entities.BeneficialOwnerType;
 import gov.medicaid.entities.EntityStructureType;
 import gov.medicaid.entities.LookupEntity;
-import gov.medicaid.entities.ProviderType;
 import gov.medicaid.entities.dto.ViewStatics;
 import gov.medicaid.services.LookupService;
 import gov.medicaid.services.util.Util;
@@ -34,12 +31,8 @@ import javax.ejb.TransactionManagement;
 import javax.ejb.TransactionManagementType;
 import javax.persistence.EntityManager;
 import javax.persistence.PersistenceContext;
-import java.util.ArrayList;
-import java.util.Arrays;
 import java.util.Collections;
-import java.util.HashSet;
 import java.util.List;
-import java.util.stream.Collectors;
 
 /**
  * Defines the UI related services.
@@ -63,30 +56,6 @@ public class LookupServiceBean implements LookupService {
      * Empty constructor.
      */
     public LookupServiceBean() {
-    }
-
-    public List<ProviderType> getAllProviderTypes() {
-        return em.createQuery(
-                "FROM ProviderType",
-                ProviderType.class
-        ).getResultList();
-    }
-
-    /**
-     * Retrieves the provider types filtered by applicant type.
-     *
-     * @param applicantType
-     *            individual or organizations
-     * @return the filtered provider types
-     */
-    public List<ProviderType> getProviderTypes(ApplicantType applicantType) {
-        return em.createQuery(
-                "FROM ProviderType WHERE applicantType = :type",
-                ProviderType.class
-        ).setParameter(
-                "type",
-                applicantType
-        ).getResultList();
     }
 
     /**
@@ -216,32 +185,5 @@ public class LookupServiceBean implements LookupService {
             results = findAllLookups(gov.medicaid.entities.BeneficialOwnerType.class);
         }
         return results;
-    }
-
-    @Override
-    public void updateProviderTypeAgreementSettings(
-            ProviderType providerType,
-            long[] agreementIds
-    ) {
-        providerType.setAgreementDocuments(
-                new HashSet<>(getAgreementDocuments(agreementIds))
-        );
-        em.merge(providerType);
-    }
-
-    private List<AgreementDocument> getAgreementDocuments(long[] agreementIds) {
-        if (agreementIds.length == 0) {
-            return new ArrayList<>();
-        } else {
-            return em.createQuery(
-                    "FROM AgreementDocument WHERE id IN :ids",
-                    AgreementDocument.class
-            ).setParameter(
-                    "ids",
-                    Arrays.stream(agreementIds)
-                            .boxed()
-                            .collect(Collectors.toList())
-            ).getResultList();
-        }
     }
 }
